fix: optimize iOS UI and reorganize transaction components
- Reduced MobileNav height to 50px (plus safe area) to match iOS standards.
- Hidden vertical scrollbars globally for a cleaner, native-app feel.
- Moved transaction sub-components from src/routes/ to src/components/transactions/
  to prevent TanStack Router warnings.
- Updated AGENTS.md with new conventions for mobile UI sizing and component organization.
